requirements.txt
asttokens==2.4.1
backcall==0.2.0
blinker @ file:///croot/blinker_1696539051175/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1698129812380/work
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
Flask @ file:///croot/flask_1716545870149/work
Flask-Cors @ file:///tmp/build/80754af9/flask_cors_1614875208036/work
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
itsdangerous @ file:///croot/itsdangerous_1716533337253/work
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml @ file:///home/conda/feedstock_root/build_artifacts/lxml_1723458263021/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-docx @ file:///home/conda/feedstock_root/build_artifacts/python-docx_1714608302440/work
pyzmq==26.2.0
requests @ file:///croot/requests_1721410876868/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.13
Werkzeug @ file:///croot/werkzeug_1716533313107/work
zipp==3.20.1
